Adds constructor in listener
