--- commonLibrary/src/main/java/com/skytask/common/Channels.java
@@ -1,7 +1,0 @@
-package com.skytask.common;
-
-public final class Channels {
-    public final static String CREATE_PRODUCT = "createProductChannel";
-    public final static String REQUEST_PRODUCT_LIST = "requestProductListChannel";
-    public final static String RESPONSE_PRODUCT_LIST = "responseProductListChannel";
-}
