+++ microservice/src/test/java/com/skytask/MicroserviceTest.java
@@ -1,8 +1,6 @@
 package com.skytask;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.skytask.common.Product;
-import com.skytask.listener.Listener;
 import com.skytask.service.ProductService;
 import org.junit.Test;
 import org.junit.runner.RunWith;
